Refined statistics reporting, adjustments to model, tree height changes

+++ src/world.rs
@@ -3,6 +3,6 @@ use giraffe::Giraffe;
 
 pub struct World {
     pub giraffes:         Vec<Giraffe>,
-    pub mutation_percent: u8,
+    pub mutation_percent: f32,
     pub tree_height:      u32
 }
